+++ tests/fixtures/get-test-route.ts
@@ -2,7 +2,7 @@ import { createWithEdgeSpec } from "src/create-with-edge-spec.js"
 import type { AxiosInstance } from "axios"
 import getPort from "@ava/get-port"
 import defaultAxios from "redaxios"
-import { createNodeServerFromRouteMap } from "src/serve/create-server-from-route-map.js"
+import { createNodeServerFromRouteMap } from "src/serve/create-node-server-from-route-map.js"
 import { EdgeSpecRouteFn } from "src/types/web-handler"
 import { ExecutionContext } from "ava"
 import { once } from "events"
